feat: implementa pacotes do pipeline geracao anuncios
