+++ ai-worker/src/main/java/com/marketinghub/pipelines/geracaoanuncios/v1/GeraAnuncioV2Pipeline.java
@@ -1,0 +1,18 @@
+package com.marketinghub.pipelines.geracaoanuncios.v1;
+
+import java.util.Set;
+
+/** Representa o núcleo declarativo do pipeline GeraAnuncio v2 no AI Worker. */
+public final class GeraAnuncioV2Pipeline {
+    public static final String TEXTO = "texto";
+    public static final String IMAGEM = "imagem";
+    private static final Set<String> ETAPAS = Set.of(TEXTO, IMAGEM);
+
+    /** Impede instanciação de classe utilitária do núcleo do pipeline. */
+    private GeraAnuncioV2Pipeline() {}
+
+    /** Informa se o código recebido representa uma etapa conhecida do pipeline. */
+    public static boolean contemEtapa(String etapa) {
+        return ETAPAS.contains(etapa);
+    }
+}
