--- backend/ads-service/src/main/java/com/marketinghub/geraanuncio/v2/imagem/service/listStageExecutions/GeraAnuncioImagemExecutionSummaryResponse.java
@@ -1,6 +1,0 @@
-package com.marketinghub.geraanuncio.v2.imagem.service.listStageExecutions;
-
-import java.time.Instant;
-
-/** Resumo de execução da etapa Imagem do GeraAnuncio v2 para relatórios. */
-public record GeraAnuncioImagemExecutionSummaryResponse(String stageExecutionId, Long experimentId, String jobId, String status, Instant updatedAt) {}
