+++ backend/ads-service/src/main/java/com/marketinghub/pipelines/facebookads/geracaoanuncios/v1/imagem/service/detailStageExecution/GeraAnuncioImagemDetailResponse.java
@@ -1,0 +1,7 @@
+package com.marketinghub.pipelines.facebookads.geracaoanuncios.v1.imagem.service.detailStageExecution;
+
+import java.time.Instant;
+import java.util.Map;
+
+/** Detalhe auditável de uma execução da etapa Imagem do GeraAnuncio v2. */
+public record GeraAnuncioImagemDetailResponse(String stageExecutionId, String jobId, String status, Instant updatedAt, Map<String, Object> reportData) {}
