+++ backend/ads-service/src/test/java/com/marketinghub/architecture/ArquiteturaTest.java
@@ -167,7 +167,7 @@ class ArquiteturaTest {
             GERALANDING_STAGE_EXECUTION_REPOSITORY_CLASS);
     private static final String MOIS_DOSSIE_V1_EXECUTOR_MODULE = "mois-sales-library-worker";
     private static final String MOIS_DOSSIE_V1_PACKAGE = "com.marketinghub.mois.dossie.v1";
-    private static final String GERAANUNCIO_V2_PACKAGE = "com.marketinghub.geraanuncio.v2";
+    private static final String GERAANUNCIO_V2_PACKAGE = "com.marketinghub.pipelines.facebookads.geracaoanuncios.v1";
     private static final Map<String, String> GERAANUNCIO_V2_STAGE_ENDPOINT_SLUGS = Map.ofEntries(
             Map.entry("texto", "texto"),
             Map.entry("imagem", "imagem"));
@@ -432,7 +432,7 @@ class ArquiteturaTest {
                     .anyMatch(method -> method.getName().equals("pending") && method.isAnnotatedWith(PostMapping.class));
             if (!hasPending) {
                 violations.add("[ARQUITETURA] [BACKEND][GeraAnuncio v2] etapa " + stagePackage
-                        + " deve expor endpoint pending canônico /api/internal/geraanuncio/v2/"
+                        + " deve expor endpoint pending canônico /api/internal/facebookads/geracaoanuncios/v1/"
                         + endpointSlug + "/stage-executions/pending");
             }
         });
@@ -442,9 +442,9 @@ class ArquiteturaTest {
     @ArchTest
     static final ArchRule geraAnuncioV2ServiceContractsMustBeRecords = classes()
             .that()
-            .resideInAPackage("com.marketinghub.geraanuncio.v2.*.service..")
+            .resideInAPackage("com.marketinghub.pipelines.facebookads.geracaoanuncios.v1.*.service..")
             .and()
-            .resideOutsideOfPackage("com.marketinghub.geraanuncio.v2.*.service")
+            .resideOutsideOfPackage("com.marketinghub.pipelines.facebookads.geracaoanuncios.v1.*.service")
             .should(beRecords())
             .because("[ARQUITETURA] [BACKEND][GeraAnuncio v2] DTOs/contratos de service devem ser records em subpacotes da operação");
