+++ docs/canonical/pipeline-operacional-canon.v1.md
@@ -111,14 +111,17 @@ O backend deve concentrar:
 7. exposição de endpoints administrativos;
 8. exposição de endpoints internos para workers.
 
-Para pipelines por etapa, o pacote deve seguir a separação funcional do GeraLanding:
+Para pipelines por etapa, o pacote deve seguir a separação funcional do GeraLanding dentro do padrão
+versionado canônico:
 
 ```text
-com.marketinghub.<dominio>.<etapa>.web
-com.marketinghub.<dominio>.<etapa>.service
-com.marketinghub.<dominio>.<etapa>.model
-com.marketinghub.<dominio>.<etapa>.repository
+com.marketinghub.pipelines.<nome-modulo>.<nome-pipeline>.v<numero-versao>.<nome-etapa>.web
+com.marketinghub.pipelines.<nome-modulo>.<nome-pipeline>.v<numero-versao>.<nome-etapa>.service
+com.marketinghub.pipelines.<nome-modulo>.<nome-pipeline>.v<numero-versao>.<nome-etapa>.model
+com.marketinghub.pipelines.<nome-modulo>.<nome-pipeline>.v<numero-versao>.<nome-etapa>.repository
 ```
+
+Exemplo backend: `com.marketinghub.pipelines.facebookads.geracaoanuncios.v1.roteiroCriativo.web`.
 
 Quando a etapa usa Worker AI, o backend deve expor o contrato mínimo:
 
@@ -135,13 +138,15 @@ A nomenclatura pode variar por domínio, mas o padrão funcional não pode varia
 
 ### 6.2 Worker AI
 
-Quando uma etapa usa OpenAI, o Worker AI deve seguir o padrão do núcleo por etapa:
+Quando uma etapa usa OpenAI, o módulo executor deve seguir o padrão canônico sem repetir o nome do módulo:
 
 ```text
-com.marketinghub.worker.openai.core.<etapa>
+com.marketinghub.pipelines.<nome-pipeline>.v<numero-versao>.<nome-etapa>
 ```
 
-Cada etapa deve possuir configuração própria, adapter de backend, construtor de prompt, validador de resposta e handler de aplicação. O core genérico não deve depender de etapas concretas.
+Exemplo no executor: `com.marketinghub.pipelines.geracaoanuncios.v1.roteiroCriativo`.
+
+Cada etapa deve possuir configuração própria, adapter de backend, construtor de prompt, validador de resposta e handler de aplicação. O núcleo genérico do executor não deve depender de etapas concretas.
 
 O worker não acessa banco. Ele busca pendências no backend pelo endpoint `pending`, faz `claim` quando o contrato da etapa exigir reserva explícita, monta prompt, chama OpenAI, valida resposta e devolve o resultado ao backend. O backend não pode substituir o worker fazendo chamada direta à OpenAI para “atalhar” uma tela ou reduzir latência; se a UI precisar de resposta rápida, o contrato deve continuar assíncrono e mostrar estado de solicitação/processamento.
