+++ docs/swagger/geraanuncio-v2-swagger.yaml
@@ -4,7 +4,7 @@ info:
   version: v2
   description: Contratos internos do pipeline GeraAnuncio v2 para geração separada de texto e imagem de anúncios pelo AI Worker.
 paths:
-  /api/internal/geraanuncio/v2/texto/stage-executions/experiments/{experimentId}/start:
+  /api/internal/facebookads/geracaoanuncios/v1/texto/stage-executions/experiments/{experimentId}/start:
     post:
       summary: Inicia execução da etapa Texto para um experimento
       description: Endpoint usado pela tela administrativa para iniciar a etapa de texto do GeraAnuncio v2.
@@ -18,14 +18,14 @@ paths:
       responses:
         '200':
           description: Solicitação registrada na etapa Texto do GeraAnuncio v2.
-  /api/internal/geraanuncio/v2/texto/stage-executions/pending:
+  /api/internal/facebookads/geracaoanuncios/v1/texto/stage-executions/pending:
     post:
       summary: Lista execuções pendentes da etapa Texto
       description: Endpoint canônico usado pelo AI Worker para consumir a fila da etapa Texto.
       responses:
         '200':
           description: Execuções pendentes retornadas com contexto e artefatos anteriores.
-  /api/internal/geraanuncio/v2/texto/stage-executions/{stageExecutionId}/prompt:
+  /api/internal/facebookads/geracaoanuncios/v1/texto/stage-executions/{stageExecutionId}/prompt:
     post:
       summary: Registra prompt enviado ao modelo na etapa Texto
       parameters:
@@ -37,7 +37,7 @@ paths:
       responses:
         '202':
           description: Prompt aceito para auditoria.
-  /api/internal/geraanuncio/v2/texto/stage-executions/{stageExecutionId}/response:
+  /api/internal/facebookads/geracaoanuncios/v1/texto/stage-executions/{stageExecutionId}/response:
     post:
       summary: Registra resposta do modelo e saída estruturada da etapa Texto
       parameters:
@@ -49,7 +49,7 @@ paths:
       responses:
         '202':
           description: Resposta aceita para auditoria e relatório.
-  /api/internal/geraanuncio/v2/imagem/stage-executions/experiments/{experimentId}/start:
+  /api/internal/facebookads/geracaoanuncios/v1/imagem/stage-executions/experiments/{experimentId}/start:
     post:
       summary: Inicia execução da etapa Imagem para um experimento
       description: Endpoint usado pela tela administrativa para iniciar a etapa de imagem do GeraAnuncio v2.
@@ -63,14 +63,14 @@ paths:
       responses:
         '200':
           description: Solicitação registrada na etapa Imagem do GeraAnuncio v2.
-  /api/internal/geraanuncio/v2/imagem/stage-executions/pending:
+  /api/internal/facebookads/geracaoanuncios/v1/imagem/stage-executions/pending:
     post:
       summary: Lista execuções pendentes da etapa Imagem
       description: Endpoint canônico usado pelo AI Worker para consumir a fila da etapa Imagem.
       responses:
         '200':
           description: Execuções pendentes retornadas com contexto e artefatos anteriores.
-  /api/internal/geraanuncio/v2/imagem/stage-executions/{stageExecutionId}/prompt:
+  /api/internal/facebookads/geracaoanuncios/v1/imagem/stage-executions/{stageExecutionId}/prompt:
     post:
       summary: Registra prompt enviado ao modelo na etapa Imagem
       parameters:
@@ -82,7 +82,7 @@ paths:
       responses:
         '202':
           description: Prompt aceito para auditoria.
-  /api/internal/geraanuncio/v2/imagem/stage-executions/{stageExecutionId}/response:
+  /api/internal/facebookads/geracaoanuncios/v1/imagem/stage-executions/{stageExecutionId}/response:
     post:
       summary: Registra resposta do modelo e saída estruturada da etapa Imagem
       parameters:
